Remove ResourceT and move to LTS-15.12
The API is using ExceptT on top of IO. This is not a good idea as
it makes the error reporting ambiguous. We should fix this.

At the same time, we were using ResourceT and that now depends on
MonadUnliftIO which was failing the build. Our REST API calls do not
return large bodies. So, I simply removed ResourceT. This allowed for
an easier transition to the newer stackage LTS version.

The library now build on the latest stackage LTS-15.12
Testing on cabal still to be done.